Implemented the Information Modals
Current implementation uses buttons with link styling and name for referencing what information to show.

Information is stored in JSON object and accesed via declared indicators in the same file.

NOTE: This solution does not yet take in to account the multiple languages.
However this should not be too hard to implement.